+++ pizzaserver-worldfile/src/main/java/io/github/willqi/pizzaserver/format/api/chunks/BedrockNetworkDiskSerializable.java
@@ -1,4 +1,4 @@
-package io.github.willqi.pizzaserver.format.api.chunks.api;
+package io.github.willqi.pizzaserver.format.api.chunks;
 
 import io.github.willqi.pizzaserver.format.BlockRuntimeMapper;
